Add the ability to create your own APIRequestLog model based on the BaseAPIRequestLog without having to add 'drf-tracking' to INSTALLED_APPS.
Do this by splitting up models and mixins into base_models and base_mixins. base_models and base_mixins are the files you can use to
create your own models and mixins, and models and mixins are the modules you can use to for the normal use case.
